Projeto_GCP/1.request_api/requirements.txt
functions-framework==3.*
requests
google-cloud-bigquery
flask
